# src/main/openapi/bus-ticket-booking-app.yml
openapi: 3.1.0
info:
  title: Bus Ticket Booking API
  description: API for managing bus ticket booking application
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearerAuth: [ ]
...
paths:
  /api/v1/wallet/recharge:
    post:
      tags:
        - Wallet Management
      summary: Recharge user wallet
      description: Allows users to recharge their wallet using various payment methods.
      operationId: RechargeWallet
      requestBody:
        description: Payment request containing payment method and transaction details
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRequest"
        required: true
      responses:
        "200":
          description: Wallet recharged successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "400":
          description: Invalid payment request
          content:
            application/json:
              schema:
                type: string
                example: Invalid payment request
        "403":
          description: Forbidden - User does not have permission to recharge wallet
          content:
            application/json:
              schema:
                type: string
                example: You do not have permission to perform this action
        "404":
          description: Coupon not found or expired
          content:
            application/json:
              schema:
                type: string
                example: "Coupon with code {couponCode} not found or expired"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                example: Internal server error occurred
...
  /api/v1/reservations:
    get:
      tags:
        - Reservation Management
      summary: Get all reservations for the authenticated user
      description: Retrieve a list of all reservations made by the authenticated user.
        Only accessible by users with USER role.
      operationId: GetUsersReservations
      responses:
        "200":
          description: Successfully retrieved user's reservations
          content:
            application/json:
              schema:
                type: array
                $ref: "#/components/schemas/ReservationResponse"
        "403":
          description: Forbidden - User does not have USER role
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReservationResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                example: Internal server error occurred
    post:
      tags:
        - Reservation Management
      summary: Create a new reservation
      description: Create a new reservation for a scheduled trip.
      operationId: CreateReservation
      requestBody:
        description: Reservation data to book the ticket
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationRequest"
        required: true
      responses:
        "201":
          description: Successfully created reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationResponse"
        "400":
          description: Bad Request - Invalid input data
          content:
            application/json:
              schema:
                type: string
                example: Invalid reservation request
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                example: Internal server error occurred
    delete:
      tags:
        - Reservation Management
      summary: Cancel a reservation
      description: Cancel an existing reservation by its ID. Only accessible by users
        with USER role.
      operationId: CancelReservation
      parameters:
        - name: reservationId
          in: query
          description: ID of the reservation to cancel
          required: true
          schema:
            type: integer
            format: int64
            example: 12345
      responses:
        "204":
          description: Successfully cancelled reservation
        "400":
          description: Bad Request - Invalid reservation ID
        "403":
          description: Forbidden - User does not have USER role
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                example: Internal server error occurred
...
components:
  schemas:
    PaymentRequest:
      type: object
      description: Request object for processing a payment
      properties:
        reservationId:
          type: integer
          format: int64
          description: Unique identifier for the reservation
          example: 12345
        method:
          type: string
          description: Payment method type
          enum:
            - COUPON
            - WALLET
        transactionType:
          type: string
          description: Type of transaction
          enum:
            - TICKET_PURCHASE
            - WALLET_RECHARGE
        couponCode:
          type: string
          description: Coupon code for the payment
          example: DISCOUNT2023
      required:
        - method
        - transactionType
    PaymentResponse:
      type: object
      description: Response object for payment processing
      properties:
        message:
          type: string
          description: Message indicating the result of the payment processing
          example: Payment successful
      required:
        - message
...
    ReservationRequest:
      type: object
      description: Request object for creating a reservation
      properties:
        scheduledTripId:
          type: integer
          format: int64
          description: Unique identifier for the scheduled trip
          example: 12345
        seatNumber:
          type: integer
          format: int32
          description: Seat number for the reservation
          example: 12
          minimum: 1
        passengerEmail:
          type: string
          description: Email address of the passenger
          example: joe@doe.com
        tariff:
          type: string
          description: Tariff details for the reservation
          enum:
            - ADULT
            - CHILD
            - STUDENT
            - SENIOR
      required:
        - scheduledTripId
        - seatNumber
        - tariff
    ReservationResponse:
      type: object
      description: Response object for reservation details
      properties:
        origin:
          type: string
          description: Origin of the trip
          example: Prague
        destination:
          type: string
          description: Destination of the trip
          example: Vienna
        departureDateTime:
          type: string
          format: date-time
          description: Departure date and time of the trip
          example: 2023-10-01T10:00:00Z
        seatNumber:
          type: integer
          format: int32
          description: Seat number for the reservation
          example: 12
        passengerEmail:
          type: string
          description: Email address of the passenger
          example: joe@doe.com
        status:
          type: string
          description: Status of the reservation
          enum:
            - RESERVED
            - PAID
            - CANCELED
            - EXPIRED
        priceCzk:
          type: number
          description: Price of the reservation in CZK
          example: 500.0
        tariff:
          type: string
          description: Tariff details for the reservation
          enum:
            - ADULT
            - CHILD
            - STUDENT
            - SENIOR
      required:
        - departureDateTime
        - destination
        - origin
        - passengerEmail
        - priceCzk
        - seatNumber
        - status
        - tariff
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT